e2e/requirements.txt
pytest==8.3.4
selenium==4.28.1
python-dotenv==1.0.1
